Redesigned addon part to allow multiple instances
The addon now exports the class object and not an instance of the class.
So it's possible to create multiple instances of the RadioHeadSerial
class. Each instances will be alive until it's destroy() function is
called. This function allows the garbage collector to destruct the
instance of the native addon part.
Also added a RadioHead patch to avoid compiler warning.

closes #3